move get/set_default_lang() to config.py
This deduplicates the code and allows for use of the function in other
modules
